* Typecasted time_t references to their appropriate types (%ld being   signed long, %lu being unsigned long).  Yes, there are some %lds,   believe it or not.

+++ modules/core/m_nick.c
@@ -20,7 +20,7 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- *   $Id: m_nick.c,v 1.62 2001/05/24 09:29:01 leeh Exp $
+ *   $Id: m_nick.c,v 1.63 2001/05/28 13:45:35 jdc Exp $
  */
 #include "handlers.h"
 #include "client.h"
@@ -877,8 +877,11 @@ nick_from_server(struct Client *client_p, struct Client *source_p, int parc,
           if (source_p->user)
             {
               add_history(source_p,1);
-              sendto_ll_serv_butone(client_p, source_p, 0, ":%s NICK %s :%lu",
-                                    parv[0], nick, source_p->tsinfo);
+              sendto_ll_serv_butone(client_p, source_p, 0,
+				":%s NICK %s :%lu",
+				parv[0],
+				nick,
+				(unsigned long) source_p->tsinfo);
             }
     }
